LayoutTests/js/dom/Promise-resolve-with-then-reject-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS value.then is called.
PASS thisValue is value
The promise is not rejected now.
PASS rejected
PASS result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
